refactor : User 디렉토리 정리 && profile 디렉토리 삭제
