--- src/main/java/com/example/Backend/domain/profiles/ProfilesController.java
@@ -1,4 +1,0 @@
-package com.example.Backend.domain.profiles;
-
-public class ProfilesController {
-}
