+++ src/main/java/com/example/Backend/domain/user/controllers/ProfileController.java
@@ -1,0 +1,4 @@
+package com.example.Backend.domain.user.controllers;
+
+public class ProfileController {
+}
